Refactor Vesilahti backend folder structure to be similar to Tampere
