--- service/src/main/kotlin/fi/vesilahti/evaka/VesilahtiIncomeCoefficientMultiplierProvider.kt
@@ -1,22 +1,0 @@
-// SPDX-FileCopyrightText: 2023 Tampere region
-//
-// SPDX-License-Identifier: LGPL-2.1-or-later
-
-package fi.vesilahti.evaka
-
-import fi.espoo.evaka.invoicing.domain.IncomeCoefficient
-import fi.espoo.evaka.invoicing.service.IncomeCoefficientMultiplierProvider
-import java.math.BigDecimal
-
-class VesilahtiIncomeCoefficientMultiplierProvider : IncomeCoefficientMultiplierProvider {
-    override fun multiplier(coefficient: IncomeCoefficient): BigDecimal =
-        when (coefficient) {
-            IncomeCoefficient.MONTHLY_WITH_HOLIDAY_BONUS -> BigDecimal("1.05") // = 12.5 / 12
-            IncomeCoefficient.MONTHLY_NO_HOLIDAY_BONUS -> BigDecimal("1.0000") // = 12 / 12
-            IncomeCoefficient.BI_WEEKLY_WITH_HOLIDAY_BONUS -> BigDecimal("2.23125") // = ???
-            IncomeCoefficient.BI_WEEKLY_NO_HOLIDAY_BONUS -> BigDecimal("2.125") // = ???
-            IncomeCoefficient.DAILY_ALLOWANCE_21_5 -> BigDecimal("21.5")
-            IncomeCoefficient.DAILY_ALLOWANCE_25 -> BigDecimal("25")
-            IncomeCoefficient.YEARLY -> BigDecimal("0.0833333") // 1 / 12
-        }
-}
