+++ service/src/main/resources/messages-vesilahti.yaml
@@ -1,0 +1,44 @@
+# SPDX-FileCopyrightText: 2023 Tampere region
+#
+# SPDX-License-Identifier: LGPL-2.1-or-later
+
+fi.vesilahti.evaka:
+  MessageProvider:
+    DECISION_HEADER:
+      fi: Vesilahden varhaiskasvatukseen liittyvät päätökset
+    DECISION_CONTENT:
+      fi: |
+        Olette hakeneet lapsellenne varhaiskasvatus- tai kerhopaikkaa Vesilahden varhaiskasvatuksesta. Päätös on luettavissa Suomi.fi-viestit -palvelusta.
+
+        Varhaiskasvatus tai kerhopaikka on hakemuksen tehneen huoltajan hyväksyttävissä Vesilahden kunnan varhaiskasvatuksen sähköisessä palvelussa osoitteessa varhaiskasvatus.vesilahti.fi. Paikkaa ei voi hyväksyä Suomi.fi-viestit -palvelussa.
+
+        Huomioittehan, että vastaus päätökseen tulee antaa kahden viikon kuluessa.
+    FEE_DECISION_HEADER:
+      fi: Vesilahden varhaiskasvatukseen liittyvät päätökset
+    FEE_DECISION_CONTENT:
+      fi: |
+        Olette saaneet päätöksen varhaiskasvatuksen asiakasmaksusta Vesilahden kunnalta. Päätös on luettavissa Suomi.fi-viestit -palvelusta.
+
+        Varhaiskasvatuksen asiakasmaksu on voimassa toistaiseksi ja perheellä on velvollisuus ilmoittaa, mikäli perheen tulot olennaisesti muuttuvat (+/- 10 %).
+
+        Lisätietoja varhaiskasvatuksen asiakasmaksuista
+
+        https://www.vesilahti.fi/varhaiskasvatusasiakasmaksut
+    VOUCHER_VALUE_DECISION_HEADER:
+      fi: Vesilahden varhaiskasvatukseen liittyvät päätökset
+    VOUCHER_VALUE_DECISION_CONTENT:
+      fi: |
+        Olette saaneet päätöksen palvelusetelin omavastuuosuudesta Vesilahden kunnalta. Päätös on luettavissa Suomi.fi-viestit -palvelusta.
+
+        Omavastuuosuus on voimassa toistaiseksi ja perheellä on velvollisuus ilmoittaa, mikäli perheen tulot olennaisesti muuttuvat (+/- 10 %).
+
+        Lisätietoja:
+
+        https://www.vesilahti.fi/varhaiskasvatusasiakasmaksut
+    ASSISTANCE_NEED_DECISION_HEADER:
+      fi: Vesilahden varhaiskasvatukseen liittyvät päätökset
+    ASSISTANCE_NEED_DECISION_CONTENT:
+      fi: |
+        Lapsellenne on tehty päätös tuen tarpeesta. Voit katsella päätöstä eVakassa.
+
+        Koska olette ottanut Suomi.fi -palvelun käyttöönne, on päätös myös luettavissa alla olevista liitteistä.
